requirements.txt
python-dotenv
pandas
numpy
pymongo==4.13
pymongo[srv]==4.13
certifi

# -e .
